move InMemoryCredentialsProvider to common for shared scenarios
