+++ src/vs/platform/credentials/common/credentials.ts
@@ -38,3 +38,47 @@ export interface ICredentialsService extends ICredentialsProvider {
 export const ICredentialsMainService = createDecorator<ICredentialsMainService>('credentialsMainService');
 
 export interface ICredentialsMainService extends ICredentialsService { }
+
+interface ISecretVault {
+	[service: string]: { [account: string]: string } | undefined;
+}
+
+export class InMemoryCredentialsProvider implements ICredentialsProvider {
+	private secretVault: ISecretVault = {};
+
+	async getPassword(service: string, account: string): Promise<string | null> {
+		return this.secretVault[service]?.[account] ?? null;
+	}
+
+	async setPassword(service: string, account: string, password: string): Promise<void> {
+		this.secretVault[service] = this.secretVault[service] ?? {};
+		this.secretVault[service]![account] = password;
+	}
+
+	async deletePassword(service: string, account: string): Promise<boolean> {
+		if (!this.secretVault[service]?.[account]) {
+			return false;
+		}
+		delete this.secretVault[service]![account];
+		if (Object.keys(this.secretVault[service]!).length === 0) {
+			delete this.secretVault[service];
+		}
+		return true;
+	}
+
+	async findPassword(service: string): Promise<string | null> {
+		return JSON.stringify(this.secretVault[service]) ?? null;
+	}
+
+	async findCredentials(service: string): Promise<Array<{ account: string, password: string }>> {
+		const credentials: { account: string, password: string }[] = [];
+		for (const account of Object.keys(this.secretVault[service] || {})) {
+			credentials.push({ account, password: this.secretVault[service]![account] });
+		}
+		return credentials;
+	}
+
+	async clear(): Promise<void> {
+		this.secretVault = {};
+	}
+}
