+++ src/vs/platform/credentials/node/credentialsMainService.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { ICredentialsChangeEvent, ICredentialsMainService } from 'vs/platform/credentials/common/credentials';
+import { ICredentialsChangeEvent, ICredentialsMainService, InMemoryCredentialsProvider } from 'vs/platform/credentials/common/credentials';
 import { Emitter } from 'vs/base/common/event';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { ILogService } from 'vs/platform/log/common/log';
@@ -150,7 +150,7 @@ export class CredentialsMainService extends Disposable implements ICredentialsMa
 
 		if (this.environmentMainService.disableKeytar) {
 			this.logService.info('Keytar is disabled. Using in-memory credential store instead.');
-			this._keytarCache = new InMemoryKeytar();
+			this._keytarCache = new InMemoryCredentialsProvider();
 			return this._keytarCache;
 		}
 
@@ -160,57 +160,19 @@ export class CredentialsMainService extends Disposable implements ICredentialsMa
 			await this._keytarCache.findCredentials('test-keytar-loads');
 		} catch (e) {
 			this.logService.warn(`Switching to using in-memory credential store instead because Keytar failed to load: ${e.message}`);
-			this._keytarCache = new InMemoryKeytar();
+			this._keytarCache = new InMemoryCredentialsProvider();
 		}
 		return this._keytarCache;
 	}
 
-	// This class doesn't implement the clear() function because we don't know
-	// what services have stored credentials. For reference, a "service" is an extension.
-	// TODO: should we clear credentials for the built-in auth extensions?
 	public clear(): Promise<void> {
+		if (this._keytarCache instanceof InMemoryCredentialsProvider) {
+			return this._keytarCache.clear();
+		}
+
+		// We don't know how to properly clear Keytar because we don't know
+		// what services have stored credentials. For reference, a "service" is an extension.
+		// TODO: should we clear credentials for the built-in auth extensions?
 		return Promise.resolve();
-	}
-}
-
-interface ISecretVault {
-	[service: string]: { [account: string]: string } | undefined;
-}
-
-// This class is used when we are unable to load Keytar properly it keeps credentials in-memory for
-// the duration of the process.
-class InMemoryKeytar implements KeytarModule {
-	private readonly secretVault: ISecretVault = {};
-
-	async getPassword(service: string, account: string): Promise<string | null> {
-		return this.secretVault[service]?.[account] ?? null;
-	}
-
-	async setPassword(service: string, account: string, password: string): Promise<void> {
-		this.secretVault[service] = this.secretVault[service] ?? {};
-		this.secretVault[service]![account] = password;
-	}
-
-	async deletePassword(service: string, account: string): Promise<boolean> {
-		if (!this.secretVault[service]?.[account]) {
-			return false;
-		}
-		delete this.secretVault[service]![account];
-		if (Object.keys(this.secretVault[service]!).length === 0) {
-			delete this.secretVault[service];
-		}
-		return true;
-	}
-
-	async findPassword(service: string): Promise<string | null> {
-		return JSON.stringify(this.secretVault[service]) ?? null;
-	}
-
-	async findCredentials(service: string): Promise<Array<{ account: string, password: string }>> {
-		const credentials: { account: string, password: string }[] = [];
-		for (const account of Object.keys(this.secretVault[service] || {})) {
-			credentials.push({ account, password: this.secretVault[service]![account] });
-		}
-		return credentials;
 	}
 }
