+++ src/vs/workbench/services/credentials/browser/credentialsService.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { ICredentialsService, ICredentialsProvider, ICredentialsChangeEvent } from 'vs/platform/credentials/common/credentials';
+import { ICredentialsService, ICredentialsProvider, ICredentialsChangeEvent, InMemoryCredentialsProvider } from 'vs/platform/credentials/common/credentials';
 import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
 import { Emitter } from 'vs/base/common/event';
 import { Disposable } from 'vs/base/common/lifecycle';
@@ -75,57 +75,5 @@ export class BrowserCredentialsService extends Disposable implements ICredential
 		if (this.credentialsProvider.clear) {
 			return this.credentialsProvider.clear();
 		}
-	}
-}
-
-interface ICredential {
-	service: string;
-	account: string;
-	password: string;
-}
-
-class InMemoryCredentialsProvider implements ICredentialsProvider {
-
-	private credentials: ICredential[] = [];
-
-	async getPassword(service: string, account: string): Promise<string | null> {
-		const credential = this.doFindPassword(service, account);
-
-		return credential ? credential.password : null;
-	}
-
-	async setPassword(service: string, account: string, password: string): Promise<void> {
-		this.deletePassword(service, account);
-		this.credentials.push({ service, account, password });
-	}
-
-	async deletePassword(service: string, account: string): Promise<boolean> {
-		const credential = this.doFindPassword(service, account);
-		if (credential) {
-			this.credentials.splice(this.credentials.indexOf(credential), 1);
-		}
-
-		return !!credential;
-	}
-
-	async findPassword(service: string): Promise<string | null> {
-		const credential = this.doFindPassword(service);
-
-		return credential ? credential.password : null;
-	}
-
-	private doFindPassword(service: string, account?: string): ICredential | undefined {
-		return this.credentials.find(credential =>
-			credential.service === service && (typeof account !== 'string' || credential.account === account));
-	}
-
-	async findCredentials(service: string): Promise<Array<{ account: string, password: string; }>> {
-		return this.credentials
-			.filter(credential => credential.service === service)
-			.map(({ account, password }) => ({ account, password }));
-	}
-
-	async clear(): Promise<void> {
-		this.credentials = [];
 	}
 }
